+++ src/CommitmentsV2/SFA.DAS.ReservationsV2.Api.Client/DependencyResolution/ReservationsClientApiConfiguration.cs
@@ -1,0 +1,12 @@
+﻿using SFA.DAS.Http.Configuration;
+
+namespace SFA.DAS.ReservationsV2.Api.Client.DependencyResolution
+{
+    /// <summary>
+    ///     Presents the reservation configuration as an IAzureActiveDirectoryClientConfiguration
+    /// </summary>
+    public class ReservationsClientApiConfiguration : Reservations.Api.Types.Configuration.ReservationsClientApiConfiguration,
+        IAzureActiveDirectoryClientConfiguration
+    {
+    }
+}
