+++ src/SFA.DAS.Commitments.Application/packages.config
@@ -33,7 +33,7 @@
   <package id="SFA.DAS.Common.Domain" version="1.4.272" targetFramework="net462" />
   <package id="SFA.DAS.Events.Api.Types" version="2.1.31" targetFramework="net462" />
   <package id="SFA.DAS.HashingService" version="1.0.179" targetFramework="net462" />
-  <package id="SFA.DAS.Http" version="3.2.3" targetFramework="net462" />
+  <package id="SFA.DAS.Http" version="1.2.4" targetFramework="net462" />
   <package id="SFA.DAS.Learners" version="1.0.0.39408" targetFramework="net45" />
   <package id="SFA.DAS.Messaging" version="3.0.0.53230" targetFramework="net45" />
   <package id="SFA.DAS.NLog.Logger" version="1.0.0.43242" targetFramework="net462" />
