+++ src/SFA.DAS.Reservations.Api.Client/DependencyResolution/ReservationsApiClientRegistry.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Net.Http;
-using SFA.DAS.Configuration;
+﻿using SFA.DAS.Reservations.Api.Client;
 using SFA.DAS.Reservations.Api.Types;
-using SFA.DAS.Reservations.Api.Types.Configuration;
 using StructureMap;
 
-namespace SFA.DAS.Reservation.Api.Client.DependencyResolution
+namespace SFA.DAS.Reservations.Api.Client.DependencyResolution
 {
     public class ReservationsApiClientRegistry : Registry
     {
