+++ src/SFA.DAS.Reservations.Api.Types/IReservationHelper.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Threading.Tasks;
+using SFA.DAS.Reservations.Api.Types.Types;
+
+namespace SFA.DAS.Reservations.Api.Types
+{
+    public interface IReservationHelper
+    {
+        Task<ReservationValidationResult> ValidateReservation(
+            ValidationReservationMessage request,
+            Func<string, object, Task<ReservationValidationResult>> call);
+    }
+}
