+++ src/SFA.DAS.Reservations.Api.Types/ReservationsHelper.cs
@@ -1,19 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using SFA.DAS.Reservations.Api.Types.Types;
 
 namespace SFA.DAS.Reservations.Api.Types
 {
-    public interface IReservationHelper
-    {
-        Task<ReservationValidationResult> ValidateReservation(
-            ValidationReservationMessage request,
-            Func<string, object, Task<ReservationValidationResult>> call);
-    }
-
     public class ReservationsHelper : IReservationHelper
     {
         public Task<ReservationValidationResult> ValidateReservation(ValidationReservationMessage request, Func<string, object, Task<ReservationValidationResult>> call)
